// wpref-frontend/src/app/api/generated/api/question.service.ts
export interface QuestionCreateRequestParams {
    title?: string;
    description?: string;
    /** Envoyer en multipart comme subject_ids&#x3D;1&amp;subject_ids&#x3D;2 (QueryDict.getlist). */
    subjectIds?: Array<number>;
    /** JSON string (liste) ex: [{\\\&quot;text\\\&quot;:\\\&quot;A\\\&quot;,\\\&quot;is_correct\\\&quot;:true},{\\\&quot;text\\\&quot;:\\\&quot;B\\\&quot;,\\\&quot;is_correct\\\&quot;:false}] */
    answerOptions?: string;
    /** JSON string (liste) ex: [{\\\&quot;kind\\\&quot;:\\\&quot;external\\\&quot;,\\\&quot;external_url\\\&quot;:\\\&quot;https://...\\\&quot;, \\\&quot;sort_order\\\&quot;:1}] */
    media?: string;
    /** Envoyer en multipart avec la même clé répétée: media_files&#x3D;&lt;file1&gt;, media_files&#x3D;&lt;file2&gt;, ... */
    mediaFiles?: Array<Blob>;
}

export interface QuestionDestroyRequestParams {
    /** ID de la question (question_id). */
    questionId: number;
}

export interface QuestionListRequestParams {
    /** Filtre exact (via DjangoFilterBackend). */
    description?: string;
    /** Recherche simple (title__icontains). */
    search?: string;
    /** Filtre exact (via DjangoFilterBackend). */
    title?: string;
}

export interface QuestionPartialUpdateRequestParams {
    /** ID de la question (question_id). */
    questionId: number;
    title?: string;
    description?: string;
    /** Envoyer en multipart comme subject_ids&#x3D;1&amp;subject_ids&#x3D;2 (QueryDict.getlist). */
    subjectIds?: Array<number>;
    /** JSON string (liste) ex: [{\\\&quot;text\\\&quot;:\\\&quot;A\\\&quot;,\\\&quot;is_correct\\\&quot;:true},{\\\&quot;text\\\&quot;:\\\&quot;B\\\&quot;,\\\&quot;is_correct\\\&quot;:false}] */
    answerOptions?: string;
    /** JSON string (liste) ex: [{\\\&quot;kind\\\&quot;:\\\&quot;external\\\&quot;,\\\&quot;external_url\\\&quot;:\\\&quot;https://...\\\&quot;, \\\&quot;sort_order\\\&quot;:1}] */
    media?: string;
    /** Envoyer en multipart avec la même clé répétée: media_files&#x3D;&lt;file1&gt;, media_files&#x3D;&lt;file2&gt;, ... */
    mediaFiles?: Array<Blob>;
}

export interface QuestionRetrieveRequestParams {
    /** ID de la question (question_id). */
    questionId: number;
}

export interface QuestionUpdateRequestParams {
    /** ID de la question (question_id). */
    questionId: number;
    title?: string;
    description?: string;
    /** Envoyer en multipart comme subject_ids&#x3D;1&amp;subject_ids&#x3D;2 (QueryDict.getlist). */
    subjectIds?: Array<number>;
    /** JSON string (liste) ex: [{\\\&quot;text\\\&quot;:\\\&quot;A\\\&quot;,\\\&quot;is_correct\\\&quot;:true},{\\\&quot;text\\\&quot;:\\\&quot;B\\\&quot;,\\\&quot;is_correct\\\&quot;:false}] */
    answerOptions?: string;
    /** JSON string (liste) ex: [{\\\&quot;kind\\\&quot;:\\\&quot;external\\\&quot;,\\\&quot;external_url\\\&quot;:\\\&quot;https://...\\\&quot;, \\\&quot;sort_order\\\&quot;:1}] */
    media?: string;
    /** Envoyer en multipart avec la même clé répétée: media_files&#x3D;&lt;file1&gt;, media_files&#x3D;&lt;file2&gt;, ... */
    mediaFiles?: Array<Blob>;
}