// wpref-frontend/src/app/api/generated/model/question.ts
/**
 * WpRef API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { QuestionMediaDto } from './question-media';
import { SubjectDto } from './subject';
import { QuestionAnswerOptionDto } from './question-answer-option';


export interface QuestionDto { 
    readonly id: number;
    title: string;
    description?: string;
    explanation?: string;
    allow_multiple_correct?: boolean;
    active?: boolean;
    is_mode_practice?: boolean;
    is_mode_exam?: boolean;
    readonly subjects: Array<SubjectDto>;
    subject_ids?: Array<number>;
    answer_options?: Array<QuestionAnswerOptionDto>;
    readonly media: Array<QuestionMediaDto>;
    readonly created_at: string;
}